(feat) First commit of the program FileCompressionUtility.
I wrote this program to save disk space. You pass it a comma delimited list of directories and it just compresses each file individually.
This can then be called from a scheduled job. This is meant for a team that works with a lot of files that need to be compressed.

I first had the configuration of directories configured in a table named Master.Appsettings, which is what the ConfigurationRepository is for,
but to make things simple I now call the class "ConfigurationWrapper" which just uses the -d command line option passed to it.

The purpose of using the library StructureMap was for me to learn dependency injection.

Example Usage:
FileCompressionUtility.exe -d "C:\Users\torre\source\CompressFiles"